[xcm] separate xcm config
